refactor(capabilities): :recycle: phase 1 — move design-state into design-system
Move framework-agnostic design concerns from lib/ into design-system/state/:
- lib/motion/motion.ts → design-system/state/motion/motion.ts (inlines local-storage to shed lib dep)
- lib/matrix-settings/matrix-settings.ts → design-system/state/matrix-rain/matrix-settings.ts (same)
- lib/command-palette/command-palette-events.ts → design-system/state/command-palette/command-palette-events.ts

Co-locate debounce privately in matrix-2d-canvas/ (only consumer), delete lib/debounce.

Move useShuffleArray out of design-system/hooks/ into lib/shuffle-array/use-shuffle-array.ts;
update content/clowns consumers to import from lib (no design-system component used the hook).

Update all importers: design-system hooks, design-system stores, features/matrix-rain-panel.
