+++ src/components/content/clowns/clowns-videos/use-clowns-videos-store.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { useShuffleArray } from "@/components/design-system/hooks/use-shuffle-array";
+import { useShuffleArray } from "@/lib/shuffle-array/use-shuffle-array";
 import { StateStore } from "@/types/component-store";
 
 const videos = [
