+++ src/components/design-system/atoms/effects/matrix-rain/matrix-rain/matrix-2d-canvas/debounce.ts
@@ -1,0 +1,14 @@
+export const debounce = <T extends (...args: Parameters<T>) => ReturnType<T>>(
+    func: T,
+    wait: number,
+): ((...args: Parameters<T>) => void) => {
+    let timeout: ReturnType<typeof setTimeout> | undefined;
+    return (...args: Parameters<T>) => {
+        if (timeout) {
+            clearTimeout(timeout);
+        }
+        timeout = setTimeout(() => {
+            func(...args);
+        }, wait);
+    };
+};
