+++ src/components/design-system/atoms/effects/matrix-rain/matrix-rain/matrix-2d-canvas/use-matrix-2d-canvas-store.ts
@@ -2,7 +2,7 @@
 
 import { MatrixRainDrawContext } from "@/types/effects/matrix-rain";
 import { useEffect, useRef, useState } from "react";
-import { debounce } from "@/lib/debounce/debounce";
+import { debounce } from "./debounce";
 import { EffectsStore } from "@/types/component-store";
 
 const matrix = "ﾊﾐﾋｰｳｼﾅﾓﾆｻﾜﾂｵﾘｱﾎﾃﾏｹﾒｴｶｷﾑﾕﾗｾﾈｽﾀﾇﾍ012345789Z:.=*+-<>".split("");
