--- src/components/design-system/hooks/use-shuffle-array.ts
@@ -1,15 +1,0 @@
-/* eslint-disable react-hooks/exhaustive-deps */
-'use client';
-
-import { useState, useEffect } from "react";
-import { shuffleArray } from "@/lib/shuffle-array/shuffle-array";
-
-export const useShuffleArray = <T>(array: T[] = [], numberOfItems: number): T[] => {
-  const [shuffled, setShuffled] = useState<T[]>(array.slice(0, numberOfItems));
-
-  useEffect(() => {
-    setShuffled(shuffleArray(array, numberOfItems));
-  }, []);
-
-  return shuffled;
-};
